RAGChatBot/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.12
botocore==1.38.12
certifi==2025.4.26
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
numpy==2.2.5
portalocker==2.10.1
protobuf==6.30.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qdrant-client==1.14.2
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
